requirements.txt
# ConnectHub - Community Business Network Platform
# Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0  # SQLite async support (default for local dev)
asyncpg==0.29.0    # PostgreSQL async support (optional, for production)
# psycopg2-binary==2.9.9  # Optional, using asyncpg for async operations

# Authentication
authlib==1.2.1
itsdangerous==2.1.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4  # Password hashing

# Forms & File Uploads
python-multipart==0.0.6
Pillow==10.1.0

# Configuration
python-dotenv==1.0.0

# Templates
jinja2==3.1.2

# Utilities
python-slugify==8.0.1
email-validator==2.1.0
pydantic[email]==2.5.0

# Rate Limiting
slowapi==0.1.9

# Email (Optional - for notifications)
sendgrid==6.10.0

# Caching (Optional)
redis==5.0.1

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production Server
gunicorn==21.2.0
